Add visual editor instructions to settings page. Update syntax highlighter throughout app to use a consistent theme.
