+++ packages/front-end/pages/experiment/[eid].tsx
@@ -52,7 +52,7 @@ import NewExperimentForm from "../../components/Experiment/NewExperimentForm";
 import MoreMenu from "../../components/Dropdown/MoreMenu";
 import InstructionsModal from "../../components/Experiment/InstructionsModal";
 import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
-import { okaidia } from "react-syntax-highlighter/dist/cjs/styles/prism";
+import { tomorrow as theme } from "react-syntax-highlighter/dist/cjs/styles/prism";
 import { useDefinitions } from "../../services/DefinitionsContext";
 import VisualCode from "../../components/Experiment/VisualCode";
 
@@ -483,7 +483,7 @@ const ExperimentPage = (): ReactElement => {
                       </div>
                       {v.description && <p>{v.description}</p>}
                       {v.value && experiment.implementation !== "visual" && (
-                        <SyntaxHighlighter language="json" style={okaidia}>
+                        <SyntaxHighlighter language="json" style={theme}>
                           {v.value}
                         </SyntaxHighlighter>
                       )}
