+++ packages/front-end/pages/settings/index.tsx
@@ -2,23 +2,16 @@ import React, { useEffect, useState } from "react";
 import useApi from "../../hooks/useApi";
 import LoadingOverlay from "../../components/LoadingOverlay";
 import { MemberRole, OrganizationSettings, useAuth } from "../../services/auth";
-import {
-  FaAngleRight,
-  FaCheck,
-  FaCreditCard,
-  FaDatabase,
-  FaKey,
-  FaPencilAlt,
-  FaUsers,
-} from "react-icons/fa";
-import Link from "next/link";
+import { FaCheck, FaCheckSquare, FaPencilAlt, FaSquare } from "react-icons/fa";
 import EditOrganizationForm from "../../components/Settings/EditOrganizationForm";
 import useForm from "../../hooks/useForm";
 import { ImplementationType } from "back-end/types/experiment";
-import Modal from "../../components/Modal";
+import { ApiKeyInterface } from "back-end/types/apikey";
+import VisualEditorInstructions from "../../components/Settings/VisualEditorInstructions";
 
 export type SettingsApiResponse = {
   status: number;
+  apiKeys: ApiKeyInterface[];
   organization?: {
     invites: {
       email: string;
@@ -53,35 +46,10 @@ export type SettingsApiResponse = {
   };
 };
 
-const links = [
-  {
-    icon: <FaUsers />,
-    display: "Team",
-    url: "/settings/team",
-  },
-  {
-    icon: <FaCreditCard />,
-    display: "Billing",
-    url: "/settings/billing",
-  },
-  {
-    icon: <FaKey />,
-    display: "API Keys",
-    url: "/settings/keys",
-  },
-  {
-    icon: <FaDatabase />,
-    display: "Data Sources",
-    url: "/settings/datasources",
-  },
-];
-
 function hasTypeChanges(
   value: {
     visual: boolean;
     code: boolean;
-    configuration: boolean;
-    custom: boolean;
   },
   types: ImplementationType[]
 ) {
@@ -95,31 +63,62 @@ function hasTypeChanges(
   return JSON.stringify(existing) !== JSON.stringify(current);
 }
 
-const SettingsPage = (): React.ReactElement => {
-  const { data, error, mutate } = useApi<SettingsApiResponse>(`/organization`);
+function hasCustomizationChanges(
+  value: {
+    customized: boolean;
+    logoPath: string;
+    primaryColor: string;
+    secondaryColor: string;
+  },
+  existing: OrganizationSettings
+) {
+  if (
+    value.customized === existing.customized &&
+    value.logoPath === existing.logoPath &&
+    value.primaryColor === existing.primaryColor &&
+    value.secondaryColor === existing.secondaryColor
+  )
+    return false;
+  return true;
+}
 
+const GeneralSettingsPage = (): React.ReactElement => {
+  const { data, error, mutate } = useApi<SettingsApiResponse>(`/organization`);
   const [editOpen, setEditOpen] = useState(false);
 
   // eslint-disable-next-line
-  const [value, _, manualUpdate] = useForm({
-    visual: false,
-    code: false,
-    configuration: false,
-    custom: false,
+  const [value, inputProps, manualUpdate] = useForm({
+    types: {
+      visual: false,
+      code: false,
+    },
+    // customization:
+    customized: false,
+    logoPath: "",
+    primaryColor: "#391c6d",
+    secondaryColor: "#50279a",
   });
-
   const { apiCall, organizations, setOrganizations, orgId } = useAuth();
 
   useEffect(() => {
-    if (!data?.organization?.settings?.implementationTypes) return;
-    const types = data.organization.settings.implementationTypes;
-    manualUpdate({
-      visual: types.includes("visual"),
-      code: types.includes("code"),
-      configuration: types.includes("configuration"),
-      custom: types.includes("custom"),
-    });
-  }, [data?.organization?.settings?.implementationTypes]);
+    if (data?.organization?.settings) {
+      const updated = { ...value };
+      const freshValues = data.organization.settings;
+      if (data?.organization?.settings?.implementationTypes) {
+        const typeArr = data.organization.settings.implementationTypes;
+        const types = {
+          visual: typeArr.includes("visual"),
+          code: typeArr.includes("code"),
+        };
+        updated.types = types;
+      }
+      updated.customized = freshValues.customized || false;
+      updated.logoPath = freshValues.logoPath || "";
+      updated.primaryColor = freshValues.primaryColor || "";
+      updated.secondaryColor = freshValues.secondaryColor || "";
+      manualUpdate(updated);
+    }
+  }, [data?.organization?.settings]);
 
   if (error) {
     return (
@@ -133,9 +132,47 @@ const SettingsPage = (): React.ReactElement => {
   }
 
   const typeChanges = hasTypeChanges(
-    value,
+    value.types,
     data?.organization?.settings?.implementationTypes || []
   );
+
+  const customizationChanges = hasCustomizationChanges(
+    value,
+    data?.organization?.settings
+  );
+
+  const ctaEnabled = typeChanges || customizationChanges;
+
+  const saveSettings = async () => {
+    const types: ImplementationType[] = value.types.visual
+      ? ["code", "visual"]
+      : ["code"];
+
+    await apiCall(`/organization`, {
+      method: "PUT",
+      body: JSON.stringify({
+        settings: {
+          implementationTypes: types,
+          customized: value.customized,
+          logoPath: value.logoPath,
+          primaryColor: value.primaryColor,
+          secondaryColor: value.secondaryColor,
+        },
+      }),
+    });
+    await mutate();
+    organizations.forEach((org) => {
+      if (org.id === orgId) {
+        org.settings = org.settings || {};
+        org.settings.implementationTypes = types;
+        org.settings.customized = value.customized;
+        org.settings.logoPath = value.logoPath;
+        org.settings.primaryColor = value.primaryColor;
+        org.settings.secondaryColor = value.secondaryColor;
+      }
+    });
+    setOrganizations(organizations);
+  };
 
   return (
     <div className="container-fluid mt-3 pagecontents">
@@ -146,168 +183,195 @@ const SettingsPage = (): React.ReactElement => {
           mutate={mutate}
         />
       )}
-      <h1>Settings</h1>
-
-      <div className="row mb-1">
-        <div className="col-auto">
-          <strong>Organization Name:</strong> {data.organization.name}{" "}
-          <a
-            href="#"
-            className="pl-1"
-            onClick={(e) => {
-              e.preventDefault();
-              setEditOpen(true);
-            }}
-          >
-            <FaPencilAlt />
-          </a>
-        </div>
-        <div className="col-auto">
-          <strong>Owner:</strong> {data.organization.ownerEmail}
-        </div>
-        {data.organization.slackTeam && (
-          <div
-            className="col-auto"
-            title={"Team: " + data.organization.slackTeam}
-          >
-            <FaCheck /> Connected to Slack
+      <h1>General Settings</h1>
+      <div className=" mb-1">
+        <div className=" bg-white p-3 border">
+          <div className="row">
+            <div className="col-sm-3">
+              <h4>Organization</h4>
+            </div>
+            <div className="col-sm-9">
+              <div className="form-group row">
+                <div className="col-sm-12">
+                  <strong>Name: </strong> {data.organization.name}{" "}
+                  <a
+                    href="#"
+                    className="pl-1"
+                    onClick={(e) => {
+                      e.preventDefault();
+                      setEditOpen(true);
+                    }}
+                  >
+                    <FaPencilAlt />
+                  </a>
+                </div>
+              </div>
+              <div className="form-group row">
+                <div className="col-sm-12">
+                  <strong>Owner:</strong> {data.organization.ownerEmail}
+                </div>
+              </div>
+              {data.organization.slackTeam && (
+                <div className="form-group row">
+                  <div
+                    className="col-sm-12"
+                    title={"Team: " + data.organization.slackTeam}
+                  >
+                    <FaCheck /> Connected to Slack
+                  </div>
+                </div>
+              )}
+            </div>
           </div>
-        )}
-      </div>
-      <div className="row mb-3">
-        <div className="col-auto">
-          <Modal
-            header="Experiment Implementation"
-            open={true}
-            inline={true}
-            autoCloseOnSubmit={false}
-            ctaEnabled={typeChanges}
-            submit={async () => {
-              if (!typeChanges) {
-                return;
-              }
-              const types = (Object.keys(value) as ImplementationType[]).filter(
-                (k) => value[k]
-              );
-              await apiCall(`/organization`, {
-                method: "PUT",
-                body: JSON.stringify({
-                  settings: {
-                    implementationTypes: types,
-                  },
-                }),
-              });
-              await mutate();
-              organizations.forEach((org) => {
-                if (org.id === orgId) {
-                  org.settings = org.settings || {};
-                  org.settings.implementationTypes = types;
-                }
-              });
-              setOrganizations(organizations);
-            }}
-            cta="Save"
-          >
-            <p>Control which implementation methods are enabled.</p>
-            <div className="form-check mb-3">
-              <input
-                type="checkbox"
-                className="form-check-input"
-                checked={value.visual}
-                onChange={(e) => {
-                  manualUpdate({
-                    visual: e.target.checked,
-                  });
-                }}
-                id="checkbox-visual"
-              />
-              <label className="form-check-label" htmlFor="checkbox-visual">
-                Visual Designer
-              </label>
-              <small className="form-text text-muted">
-                <strong>Requires our Browser SDK.</strong> Change html/css
-                without writing code.
-              </small>
+          <div className="divider border-bottom mb-3 mt-2"></div>
+          <div className="row">
+            <div className="col-sm-3">Visual Editor</div>
+            <div className="col-sm-9 pb-3">
+              <p>
+                The Visual Editor allows non-technical users to create and start
+                experiments in production without writing any code.
+              </p>
+              <div>
+                <button
+                  className={
+                    value.types.visual
+                      ? "btn btn-sm btn-primary active"
+                      : "btn btn-sm btn-secondary"
+                  }
+                  onClick={(e) => {
+                    e.preventDefault();
+                    manualUpdate({
+                      types: {
+                        code: true,
+                        visual: !value.types.visual,
+                      },
+                    });
+                  }}
+                >
+                  {value.types.visual ? <FaCheckSquare /> : <FaSquare />}{" "}
+                  {value.types.visual ? "Enabled" : "Disabled"}
+                </button>
+              </div>
+              {value.types.visual &&
+                data.organization.settings?.implementationTypes?.includes(
+                  "visual"
+                ) && (
+                  <div className="bg-light p-3 my-3 border rounded">
+                    <h5 className="font-weight-bold">Setup Instructions</h5>
+                    <VisualEditorInstructions
+                      apiKeys={data.apiKeys}
+                      mutate={mutate}
+                    />
+                  </div>
+                )}
             </div>
-            <div className="form-check mb-3">
-              <input
-                type="checkbox"
-                className="form-check-input"
-                checked={value.code}
-                onChange={(e) => {
-                  manualUpdate({
-                    code: e.target.checked,
-                  });
-                }}
-                id="checkbox-code"
-              />
-              <label className="form-check-label" htmlFor="checkbox-code">
-                Code
-              </label>
-              <small className="form-text text-muted">
-                <strong>Requires our Browser, NodeJS, PHP, or Ruby SDK</strong>
-              </small>
+          </div>
+          <div className="divider border-bottom mb-3 mt-2"></div>
+          <div className="row">
+            <div className="col-sm-3">
+              <h4>Customization</h4>
             </div>
-            <div className="form-check mb-3">
-              <input
-                type="checkbox"
-                className="form-check-input"
-                checked={value.configuration}
-                onChange={(e) => {
-                  manualUpdate({
-                    configuration: e.target.checked,
-                  });
-                }}
-                id="checkbox-configuration"
-              />
-              <label
-                className="form-check-label"
-                htmlFor="checkbox-configuration"
-              >
-                Feature Flags
-              </label>
-              <small className="form-text text-muted">
-                <strong>Requires our Browser, NodeJS, PHP, or Ruby SDK</strong>
-              </small>
-            </div>
-            <div className="form-check">
-              <input
-                type="checkbox"
-                className="form-check-input"
-                checked={value.custom}
-                onChange={(e) => {
-                  manualUpdate({
-                    custom: e.target.checked,
-                  });
-                }}
-                id="checkbox-custom"
-              />
-              <label className="form-check-label" htmlFor="checkbox-custom">
-                Custom
-              </label>
-              <small className="form-text text-muted">
-                No requirements. Leaves implementation entirely up to you.
-              </small>
-            </div>
-          </Modal>
-        </div>
-        <div className="col-auto pt-4">
-          <h3>Additional Settings</h3>
-          <div className="list-group">
-            {links.map((l, i) => (
-              <Link key={i} href={l.url}>
-                <a className="list-group-item list-group-item-action">
-                  <div className="row">
-                    <div className="col-auto">{l.icon}</div>
-                    <div className="col-auto">{l.display}</div>
-                    <div style={{ flex: 1 }} />
-                    <div className="col-auto">
-                      <FaAngleRight />
+            <div className="col-sm-9">
+              <div className="form-group row">
+                <div className="col-auto ">
+                  <div className="form-check">
+                    <input
+                      type="checkbox"
+                      className="form-check-input "
+                      checked={value.customized}
+                      onChange={(e) => {
+                        manualUpdate({
+                          customized: e.target.checked,
+                        });
+                      }}
+                      id="checkbox-customized"
+                    />
+
+                    <label
+                      htmlFor="checkbox-customized"
+                      className="form-check-label"
+                    >
+                      Enable customization
+                    </label>
+                  </div>
+                </div>
+                <div className="col-sm-9"></div>
+              </div>
+              {value.customized && (
+                <>
+                  <div className="form-group row">
+                    <div className="col-sm-3 col-form-label">
+                      <label htmlFor="customlogo">Custom logo</label>
+                    </div>
+                    <div className="col-sm-9">
+                      <input
+                        type="text"
+                        className="form-control"
+                        id="customlogo"
+                        placeholder="/path/to/logo.png"
+                        {...inputProps.logoPath}
+                      />
+                      <p>
+                        <small className="text-muted">
+                          Logo will be scaled to fit 225 x 46
+                        </small>
+                      </p>
                     </div>
                   </div>
-                </a>
-              </Link>
-            ))}
+                  <div className="form-group row">
+                    <div className="col-sm-3 col-form-label">
+                      <label htmlFor="formGroupExampleInput">
+                        Primary Color
+                      </label>
+                    </div>
+                    <div className="col-sm-9">
+                      <input
+                        className="form-control"
+                        type="color"
+                        id="primarycolor"
+                        name="primarycolor"
+                        {...inputProps.primaryColor}
+                      />
+                    </div>
+                  </div>
+                  <div className="form-group row">
+                    <div className="col-sm-3 col-form-label">
+                      <label htmlFor="formGroupExampleInput">
+                        Secondary Color
+                      </label>
+                    </div>
+                    <div className="col-sm-9">
+                      <input
+                        className="form-control"
+                        type="color"
+                        id="secondarycolor"
+                        name="secondarycolor"
+                        {...inputProps.secondaryColor}
+                      />
+                    </div>
+                  </div>
+                </>
+              )}
+            </div>
+          </div>
+          <div className="divider border-bottom mb-3 mt-3"></div>
+          <div className="row">
+            <div className="col-12">
+              <div className=" d-flex flex-row-reverse">
+                <button
+                  className={`btn btn-${ctaEnabled ? "primary" : "secondary"}`}
+                  type="submit"
+                  disabled={!ctaEnabled}
+                  onClick={async (e) => {
+                    e.preventDefault();
+                    if (!ctaEnabled) return;
+                    saveSettings();
+                  }}
+                >
+                  Save
+                </button>
+              </div>
+            </div>
           </div>
         </div>
       </div>
@@ -315,4 +379,4 @@ const SettingsPage = (): React.ReactElement => {
   );
 };
 
-export default SettingsPage;
+export default GeneralSettingsPage;
