+++ packages/front-end/styles/global.scss
@@ -10,9 +10,6 @@ $h6-font-size: $font-size-base !default;
 
 @import "bootstrap/scss/bootstrap.scss";
 @import "react-bootstrap-typeahead/css/Typeahead.scss";
-
-@import "prismjs/themes/prism-okaidia.css";
-
 @import url("https://fonts.googleapis.com/css?family=Roboto:400,700,300");
 
 html,
